Code cleanup and code splitting for simplicity
