+++ src/components/card/CardData.js
@@ -18,11 +18,10 @@ const CardData = ( { cardID, cardName, cardColor, cardEffect, cardAttribute, mon
 			Card Information
 		</Typography>
 
-		<Suspense>
-			<CardImageRounded
-				cardID={cardID}
-				/>
-		</Suspense>
+
+		<CardImageRounded
+			cardID={cardID}
+			/>
 
 		<Suspense fallback={<Skeleton width={'100%'} height={150} />} >
 			<YGOCard
